requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==45.0.3
daphne==4.2.0
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
msgpack==1.1.0
packaging==25.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==6.2.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
